twitter/requirements.txt
argcomplete==1.8.2
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.5.3
bleach==2.1.3
boto3==1.7.64
botocore==1.10.64
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==2.3.0
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
decorator==4.2.1
Django==1.11
django-nested-inline==0.3.7
docutils==0.14
docx2txt==0.6
EbookLib==0.15
entrypoints==0.2.3
evdev==0.7.0
gunicorn==19.9.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.2.3
MarkupSafe==1.0
mistune==0.8.3
mysql-connector==2.1.6
mysql-connector-python-rf==2.2.2
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.14.2
pandas==0.22.0
pandocfilters==1.4.2
parsel==1.5.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
pocketsphinx==0.1.3
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
pycrypto==2.6.1
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
PyMouse==1.0
PyMySQL==0.9.2
pynput==1.3.10
PyOpenGL==3.1.0
pyOpenSSL==18.0.0
python-dateutil==2.7.2
python-docx==0.8.6
python-pptx==0.6.5
python-xlib==0.21
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
queuelib==1.5.0
requests==2.18.4
s3transfer==0.1.13
Scrapy==1.5.1
selenium==3.11.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.10.0
smime==0.0.3
SpeechRecognition==3.6.3
SQLAlchemy==1.2.10
terminado==0.8.1
testpath==0.3.1
textract==1.6.1
tornado==5.0.2
traitlets==4.3.2
Twisted==18.7.0
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.0
xlrd==1.0.0
XlsxWriter==1.0.5
zope.interface==4.5.0
